requirements.txt
cerberus==1.2
pytest==4.3.1
factory_boy==2.11.1
SQLAlchemy==1.3.0
sqlalchemy-utils===0.33.11
psycopg2===2.7.3.2
alembic===1.0.8
graphene===2.1.3
PyYAML==5.1
python-i18n[YAML]==0.3.1
